Extract Builder as a top-level class
